feat: remove use of inspector API in favor of SWC transform
