+++ package.json
@@ -81,7 +81,7 @@
       "/node_modules/"
     ],
     "transform": {
-      "^.+\\.(t|j)sx?$": "./lib/jest"
+      "^.+\\.(t|j)sx?$": "@swc/jest"
     },
     "testMatch": [
       "<rootDir>/src/**/__tests__/**/*.ts?(x)",
