+++ src/global-singleton.ts
@@ -1,0 +1,3 @@
+export function globalSingleton<T>(key: symbol, create: () => T): T {
+  return ((global as any)[key] = (global as any)[key] ?? create());
+}
