+++ src/index.ts
@@ -1,2 +1,4 @@
+// ensure that the Function.prototype.bind function has been patched
+import "./bound-function";
+
 export { serializeFunction, SerializeFunctionProps } from "./serialize-closure";
-export * from "./globals";
