--- src/jest.ts
@@ -1,8 +1,0 @@
-import "./globals";
-
-// @ts-ignore
-import { createTransformer } from "@swc/jest";
-
-exports = { createTransformer };
-
-console.log(exports);
